restructured components to folders
