+++ src/App.tsx
@@ -1,41 +1,7 @@
 // @ts-nocheck  // removes annoying parameter any type messages for file
-import ReactImg from "./assets/react-core-concepts.png"; // import images from assets folder
 import { CORE_CONCEPTS } from "./data";
-
-const reactDescriptions = ["Fundamental", "Crucial", "Core"];
-
-// @ts-ignore // disables annoying parameter any type message for function
-function genRandomInt(max) {
-  // random number generator
-  return Math.floor(Math.random() * (max + 1));
-}
-
-function Header() {
-  const description = reactDescriptions[genRandomInt(2)];
-
-  return (
-    <header>
-      <img src={ReactImg} alt="Stylized atom" />
-      <h1>React Essentials</h1>
-      <p>
-        {description} React concepts you will need for almost any app you are
-        going to build!
-      </p>
-    </header>
-  );
-}
-
-function CoreConcept({image, title, description}) {
-  // object converted via destructuring
-  return (
-    // equivalent to parameter props.title etc.
-    <li>
-      <img src={image} alt={title} />
-      <h3>{title}</h3>
-      <p>{description}</p>
-    </li>
-  );
-}
+import Header from './components/Header'
+import CoreConcept from './components/CoreConcept'
 
 function App() {
   return (
